# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "receipt-system",
    "description": "AI\u99c6\u52d5\u578b\u78ba\u5b9a\u7533\u544a\u7528\u30ec\u30b7\u30fc\u30c8\u7ba1\u7406\u30b7\u30b9\u30c6\u30e0 - Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/files/{path}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u753b\u50cf\u53d6\u5f97",
        "description": "\u627f\u8a8d\u6e08\u307f\u30fb\u672a\u89e3\u6790\u30fb\u5931\u6557\u30d5\u30a9\u30eb\u30c0\u304b\u3089\u753b\u50cf\u3092\u914d\u4fe1\u3057\u307e\u3059\u3002",
        "operationId": "serve_file_api_v1_files__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts": {
      "get": {
        "tags": [
          "Receipts"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u4e00\u89a7\u53d6\u5f97",
        "description": "\u30ec\u30b7\u30fc\u30c8\u306e\u4e00\u89a7\u3092\u53d6\u5f97\u3057\u307e\u3059\uff08\u30d5\u30a3\u30eb\u30bf\u30fb\u30da\u30fc\u30b8\u30f3\u30b0\u5bfe\u5fdc\uff09\u3002",
        "operationId": "list_receipts_api_v1_receipts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{receipt_id}": {
      "get": {
        "tags": [
          "Receipts"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u8a73\u7d30\u53d6\u5f97",
        "description": "\u6307\u5b9a\u3055\u308c\u305f\u30ec\u30b7\u30fc\u30c8\u306e\u8a73\u7d30\u60c5\u5831\u3092\u53d6\u5f97\u3057\u307e\u3059\u3002",
        "operationId": "get_receipt_api_v1_receipts__receipt_id__get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Receipts"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u524a\u9664",
        "description": "\u6307\u5b9a\u3055\u308c\u305f\u30ec\u30b7\u30fc\u30c8\u3092\u524a\u9664\u3057\u307e\u3059\u3002",
        "operationId": "delete_receipt_api_v1_receipts__receipt_id__delete",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{receipt_id}/image": {
      "get": {
        "tags": [
          "Receipts"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u753b\u50cf\u53d6\u5f97",
        "description": "\u6307\u5b9a\u3055\u308c\u305f\u30ec\u30b7\u30fc\u30c8\u306e\u753b\u50cfURL\u3092\u53d6\u5f97\u3057\u307e\u3059\u3002",
        "operationId": "get_receipt_image_api_v1_receipts__receipt_id__image_get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptImageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{receipt_id}/reprocess": {
      "post": {
        "tags": [
          "Receipts"
        ],
        "summary": "\u30ec\u30b7\u30fc\u30c8\u518d\u89e3\u6790",
        "description": "\u6307\u5b9a\u3055\u308c\u305f\u30ec\u30b7\u30fc\u30c8\u3092\u518d\u89e3\u6790\u3057\u307e\u3059\uff08OCR\u30fbAI\u89e3\u6790\u306e\u518d\u5b9f\u884c\uff09\u3002",
        "operationId": "reprocess_receipt_api_v1_receipts__receipt_id__reprocess_post",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Receipt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptReprocessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptReprocessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReceiptDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ReceiptDeleteResponse",
        "description": "Receipt delete response schema."
      },
      "ReceiptDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "stored_filename": {
            "type": "string",
            "title": "Stored Filename"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "image_hash": {
            "type": "string",
            "title": "Image Hash"
          },
          "ocr_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Text"
          },
          "ocr_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Confidence"
          },
          "ocr_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Language"
          },
          "receipt_date": {
            "type": "string",
            "format": "date-time",
            "title": "Receipt Date"
          },
          "store_name": {
            "type": "string",
            "title": "Store Name"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "tax_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "category_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Confidence"
          },
          "ai_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Comment"
          },
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Message"
          },
          "processing_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Started At"
          },
          "processing_completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Completed At"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "stored_filename",
          "file_path",
          "file_size",
          "mime_type",
          "image_hash",
          "receipt_date",
          "store_name",
          "total_amount",
          "currency",
          "status",
          "retry_count",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "ReceiptDetailResponse",
        "description": "Receipt detail response schema."
      },
      "ReceiptImageResponse": {
        "properties": {
          "receipt_id": {
            "type": "integer",
            "title": "Receipt Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          }
        },
        "type": "object",
        "required": [
          "receipt_id",
          "image_url",
          "mime_type",
          "file_size"
        ],
        "title": "ReceiptImageResponse",
        "description": "Receipt image response schema."
      },
      "ReceiptListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReceiptResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ReceiptListResponse",
        "description": "Receipt list response schema."
      },
...
      "ReceiptReprocessRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "ReceiptReprocessRequest",
        "description": "Receipt reprocess request schema."
      },
      "ReceiptReprocessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "receipt_id": {
            "type": "integer",
            "title": "Receipt Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "receipt_id"
        ],
        "title": "ReceiptReprocessResponse",
        "description": "Receipt reprocess response schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}